requirements.txt
#kernel: python3
ipykernel>=6.29.5

# Excel processing
openpyxl>=3.1.2
xlrd>=2.0.1
pandas>=2.0.3

# PDF processing
PyMuPDF>=1.23.7  # Used for PDF reading and rendering to images (no external dependencies)

# Word document processing
python-docx>=0.8.11
docx2python>=1.0.0

# PowerPoint processing
python-pptx>=0.6.21

# Windows-specific dependencies (only installed on Windows)
pywin32>=306; sys_platform == 'win32'

# OCR support
pytesseract>=0.3.10
# Note: tesseract-ocr must be installed separately on the system

# Encoding detection
chardet>=5.2.0

# Core utilities
typing-extensions>=4.8.0

# Data analysis and machine learning
scikit-learn>=1.3.0
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
seqeval>=1.2.2

# Deep learning frameworks
torch>=2.0.0,<2.2.0
transformers>=4.35.0,<5.0.0
datasets>=2.14.0

# French language models
sentencepiece>=0.1.99

# Model evaluation and utilities
tqdm>=4.65.0
tensorboard>=2.13.0
sklearn-crfsuite>=0.3.6

# Document processing
pypdf>=3.0.0
pdfminer.six>=20221105
Pillow>=10.0.0
nltk>=3.8.0
wordcloud>=1.9.0
spacy>=3.7.0,<4.0.0
# SpaCy French model will be downloaded separately

sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
